refactor: unify terminal blocked comment helper across heartbeat and pipeline
